+++ example/lib/main_select_from_listenable.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:implicit_navigator/implicit_navigator.dart';
 
 /// This is a version of the Flutter Navigator 2.0 tutorial (sans routing logic)
-/// rebuilt using Implicit Navigator!
+/// rebuilt using Implicit Navigator's `selectFromListenable` method!
 ///
 /// See:
 /// * https://medium.com/flutter/learning-flutters-new-navigation-and-routing-system-7c9068155ade
@@ -21,13 +21,8 @@ class Book {
   final String author;
 }
 
-class BooksApp extends StatefulWidget {
-  @override
-  State<StatefulWidget> createState() => _BooksAppState();
-}
-
-class _BooksAppState extends State<BooksApp> {
-  Book? _selectedBook;
+class BooksApp extends StatelessWidget {
+  final SelectedBook _selectedBook = SelectedBook(null);
 
   static const List<Book> books = [
     Book('Left Hand of Darkness', 'Ursula K. Le Guin'),
@@ -39,13 +34,14 @@ class _BooksAppState extends State<BooksApp> {
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Books App',
-      home: ImplicitNavigator<Book?>(
-        value: _selectedBook,
+      home: ImplicitNavigator.selectFromListenable<SelectedBook, Book?>(
+        listenable: _selectedBook,
+        selector: () => _selectedBook._book,
         builder: (context, book, animation, secondaryAnimation) {
           if (book == null) {
             return BooksListScreen(
               books: books,
-              onTapped: _handleBookTapped,
+              onTapped: (newBook) => _selectedBook.book = newBook,
             );
           }
           return BookDetailsScreen(
@@ -53,16 +49,10 @@ class _BooksAppState extends State<BooksApp> {
             book: book,
           );
         },
+        onPop: (poppedBook, bookAfterPop) => _selectedBook.book = bookAfterPop,
         transitionsBuilder: ImplicitNavigator.materialRouteTransitionsBuilder,
-        onPop: (poppedBook, currentBook) => _selectedBook = currentBook,
       ),
     );
-  }
-
-  void _handleBookTapped(Book book) {
-    setState(() {
-      _selectedBook = book;
-    });
   }
 }
 
@@ -116,5 +106,18 @@ class BookDetailsScreen extends StatelessWidget {
         ),
       ),
     );
+  }
+}
+
+class SelectedBook extends ChangeNotifier {
+  SelectedBook(this._book);
+
+  Book? _book;
+
+  Book? get book => _book;
+
+  set book(Book? newBook) {
+    _book = newBook;
+    notifyListeners();
   }
 }